+++ src/app/services/dashboard/dashboardKeyBindings.js
@@ -18,8 +18,8 @@ function(angular, $) {
         keyboardManager.unbind('ctrl+s');
         keyboardManager.unbind('ctrl+r');
         keyboardManager.unbind('ctrl+z');
-        keyboardManager.unbind('esc');
       });
+      keyboardManager.unbind('esc');
 
       keyboardManager.bind('ctrl+f', function(evt) {
         scope.emitAppEvent('open-search', evt);